Convert remaining non-nesting Sass functionality to standard CSS
This converts any remaining variables, all mixins except those noted as a provision for OpenClinica), and functions to their CSS equivalent (or a suitable replacement where necessary). It also removes a variety of IE-specific styles, as IE is no longer supported and it reduced the surface area of this change.

Though noisy, the vast majority of the change (apart from the hopefully obvious variable conversions) is generally either:

- eliminating obsolete mixins (i.e. they were used to handle cross-browser support that is no longer needed)
- moving things from more substantive mixins inline
- adding generalized classes for cases where those more substantive mixins were used in multiple places (this was generally isolated to widgets, where they are already templating or toggling classes, so there should be negligible performance impact)
- adapting some themed data: URL images to use masks (CSS variables cannot be applied to data: URLs)
- simplifying some styles in cases where the opportunity to rely on CSS variable cascading behavior was obvious